Removed a significant portion of unused/ needed code.
Removed references to admin account as a hotspot should not be able to log in as an admin to the iroha instance.

Added to the main file's listen_push_data to make it collect packets properly and push them to the blockchain.

cleaned up the output by adding a debug mode and modifying any print statements.